modification by koreth (http://www.dgrin.com/member.php?u=46644)
I made another tweak to the script. Now if --captionsfromdescription is
specified, it will try to use the description, then the summary, then
the title, then the filename (unless --nofilenamecaptions is specified)
until it finds something. Also, if --nofilenamecaptions is specified, it
now checks for filename titles that got saved to the Gallery2 database
and strips them off (previously that option would only avoid creating
*new* filename captions but would use existing ones.)